move fft_phase_1d to util.py
