+++ diffmap/__init__.py
@@ -1,2 +1,3 @@
 from .diffmap import *
 from .codiffmap import *
+from .util import *
